proj_slam/SLAM-Former/requirements.txt
torch==2.9.1
torchvision==0.24.1
numpy==2.2.6
Pillow
huggingface_hub
safetensors
roma
gradio
matplotlib
tqdm
opencv-python
scipy
einops
trimesh
tensorboard
pyglet<2
viser
gradio
lpips
hydra-core
h5py
accelerate
transformers
scikit-learn
gsplat
evo
open3d
rerun-sdk
kornia
moviepy==1.0.3